@@ -1,25 +1,24 @@
-name: flutter_http
-description: A new Flutter application.
-
+name: flutter_dio
+description: "A new Flutter project to demonstrate the Dio package usage."
 publish_to: 'none'
-
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-
+  sdk: ^3.7.0
 dependencies:
   flutter:
     sdk: flutter
-  json_annotation: ^4.4.0
-  dio: ^4.0.6
+  json_annotation: ^4.9.0
   weather_icons: ^3.0.0
+  dio: ^5.8.0+1
+  cupertino_icons: ^1.0.8
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^2.1.8
-  json_serializable: ^6.1.5
+  build_runner: ^2.4.15
+  json_serializable: ^6.9.4
+  flutter_lints: ^5.0.0
 
 flutter:
   uses-material-design: true
