@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_http/list_page.dart';
+import 'package:flutter_dio/list_page.dart';
 
 void main() {
   runApp(MyApp());
@@ -8,14 +8,13 @@ void main() {
 final scaffoldMessengerKey = GlobalKey<ScaffoldMessengerState>();
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
       scaffoldMessengerKey: scaffoldMessengerKey,
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
       home: ListPageWidget(),
     );
   }
